Fixed signed/unsigned warnings and fixed tests
darcs-hash:20061018003507-fafb9-90e335a8dcfe9c9358c4ef1cfa6c0f2a63547fa7.gz

+++ perfMedian.cxx
@@ -17,7 +17,7 @@ int main(int, char * argv[])
   typedef unsigned char PType;
   typedef itk::Image< PType, dim > IType;
 
-  int repeats = atoi(argv[1]);
+  unsigned repeats = (unsigned)atoi(argv[1]);
 
   // set up the radius
   typedef std::vector<int> IVec;
