+++ test2DShortHistMean.cxx
@@ -14,7 +14,7 @@ int main(int, char * argv[])
   typedef short PType;
   typedef itk::Image< PType, dim > IType;
 
-  int repeats = atoi(argv[1]);
+  unsigned repeats = (unsigned)atoi(argv[1]);
   itk::TimeProbe HTime, TTime;
 
   typedef itk::ImageFileReader< IType > ReaderType;
